requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
bleach==6.0.0
boto3==1.28.3
botocore==1.31.3
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.5
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1686613057838/work
cycler==0.11.0
dacite==1.7.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
-e git+https://github.com/facebookresearch/detectron2.git@47d8ed7a1bd999719825f2e8a11d5b2828bc02c4#egg=detectron2
dill==0.3.6
dnspython==2.4.0
docker-pycreds==0.4.0
eventlet==0.33.3
executing==1.2.0
fastjsonschema==2.17.1
fiftyone==0.21.4
fiftyone-brain==0.13.0
fiftyone-db==0.4.0
fiftyone-desktop==0.28.2
filelock @ file:///work/ci_py311/filelock_1676823024085/work
fonttools==4.41.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.32
glob2==0.7
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
google-auth==2.22.0
google-auth-oauthlib==1.0.0
graphql-core==3.2.3
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hydra-core==1.3.2
hypercorn==0.14.4
hyperframe==6.0.1
idna @ file:///work/ci_py311/idna_1676822698822/work
imageio==2.31.1
inflate64==0.3.1
iopath==0.1.9
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
Jinja2 @ file:///work/ci_py311/jinja2_1676823587943/work
jmespath==1.0.1
joblib==1.3.1
jsonlines==3.1.0
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kaleido==0.2.1
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
Markdown==3.4.3
MarkupSafe @ file:///work/ci_py311/markupsafe_1676823507015/work
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mongoengine==0.24.2
motor==3.2.0
mpmath==1.2.1
multidict==6.0.4
multivolumefile==0.2.3
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx @ file:///croot/networkx_1678964333703/work
notebook==6.5.4
notebook_shim==0.2.3
numpy @ file:///croot/numpy_and_numpy_base_1687466208449/work
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
openslide-python==1.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.9.1
plotly==5.15.0
portalocker==2.7.0
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.18.0
Pygments==2.15.1
pymongo==4.4.1
pyOpenSSL @ file:///croot/pyopenssl_1678965284384/work
pyparsing==3.0.9
pyppmd==1.0.0
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
pyzstd==0.15.9
qtconsole==5.4.3
QtPy==2.3.1
rarfile==4.0
referencing==0.29.1
regex==2023.6.3
requests @ file:///croot/requests_1682607517574/work
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.11
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
sentry-sdk==1.28.1
setproctitle==1.3.2
shapely @ file:///croot/shapely_1680636282912/work
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.6.2
starlette==0.26.1
strawberry-graphql==0.138.1
sympy @ file:///work/ci_py311_2/sympy_1679339311852/work
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.2
tinycss2==1.2.1
torch==2.0.1
torchmetrics==1.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1686601443660/work
tzdata==2023.3
tzlocal==5.0.1
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1686163155763/work
voxel51-eta==0.10.0
wandb==0.15.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
yacs==0.1.8
yarl==1.9.2
